% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\name{Collection}
\alias{Collection}
\title{Collection Class}
\description{
Collection Class

Collection Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Id of the collection}

\item{\code{title}}{Collection title}

\item{\code{description}}{Short description of the collection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Collection$new()}}
\item \href{#method-setCollection}{\code{Collection$setCollection()}}
\item \href{#method-getCollection}{\code{Collection$getCollection()}}
\item \href{#method-setMetadata}{\code{Collection$setMetadata()}}
\item \href{#method-getMetadata}{\code{Collection$getMetadata()}}
\item \href{#method-collectionInfo}{\code{Collection$collectionInfo()}}
\item \href{#method-collectionInfoExtended}{\code{Collection$collectionInfoExtended()}}
\item \href{#method-clone}{\code{Collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$new(id = NA, title = NA, description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the collection}

\item{\code{title}}{Collection title}

\item{\code{description}}{Short description of the collection}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCollection"></a>}}
\if{latex}{\out{\hypertarget{method-setCollection}{}}}
\subsection{Method \code{setCollection()}}{
Add image collection to the collection class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$setCollection(ImageCollection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageCollection}}{Collection of class 'image collection'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCollection"></a>}}
\if{latex}{\out{\hypertarget{method-getCollection}{}}}
\subsection{Method \code{getCollection()}}{
Get assigned image collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$getCollection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
image collection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-setMetadata}{}}}
\subsection{Method \code{setMetadata()}}{
add extent and bands to the metadata of the collection object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$setMetadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-getMetadata}{}}}
\subsection{Method \code{getMetadata()}}{
Get metadata of the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$getMetadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
metadata of the collection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collectionInfo"></a>}}
\if{latex}{\out{\hypertarget{method-collectionInfo}{}}}
\subsection{Method \code{collectionInfo()}}{
List metadata for the collection handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$collectionInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collectionInfoExtended"></a>}}
\if{latex}{\out{\hypertarget{method-collectionInfoExtended}{}}}
\subsection{Method \code{collectionInfoExtended()}}{
List extended metadata for the collection handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$collectionInfoExtended()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
