% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutableProcess-class.R
\name{ExecutableProcess}
\alias{ExecutableProcess}
\title{Executable process}
\description{
Executable process

Executable process
}
\section{Super class}{
\code{\link[openEo.gdalcubes:Process]{openEo.gdalcubes::Process}} -> \code{ExecutableProcess}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job}}{Attached job to the process}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExecutableProcess$new()}}
\item \href{#method-execute}{\code{ExecutableProcess$execute()}}
\item \href{#method-clone}{\code{ExecutableProcess$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="openEo.gdalcubes" data-topic="Process" data-id="processInfo">}\href{../../openEo.gdalcubes/html/Process.html#method-processInfo}{\code{openEo.gdalcubes::Process$processInfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="openEo.gdalcubes" data-topic="Process" data-id="setParameter">}\href{../../openEo.gdalcubes/html/Process.html#method-setParameter}{\code{openEo.gdalcubes::Process$setParameter()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize executable process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$new(
  id = NA,
  description = NA,
  parameters = NA,
  operation = NA,
  process = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id or name of the proces}

\item{\code{description}}{Shortly what the process does}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{operation}}{Function that executes the process}

\item{\code{process}}{Processes which will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Run the operation including a generated list of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$execute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
