% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{load_collectionB}
\alias{load_collectionB}
\title{load collectionB}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 11.
}
\usage{
load_collectionB
}
\description{
load collectionB
}
\keyword{datasets}
