% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-class.R
\name{Process}
\alias{Process}
\title{Process}
\description{
Process

Process
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Id or name of the process}

\item{\code{summary}}{Shortly what the process does}

\item{\code{description}}{Extended description of the process}

\item{\code{categories}}{To which categories the process belongs}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{returns}}{The output of the process}

\item{\code{operation}}{Function that executes the process}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Process$new()}}
\item \href{#method-processInfo}{\code{Process$processInfo()}}
\item \href{#method-clone}{\code{Process$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$new(
  id = NA,
  summary = NA,
  description = NA,
  categories = NA,
  parameters = NA,
  returns = NA,
  operation = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id or name of the process}

\item{\code{summary}}{Shortly what the process does}

\item{\code{description}}{Extended description of the process}

\item{\code{categories}}{To which categories the process belongs}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{returns}}{The output of the process}

\item{\code{operation}}{Function that executes the process}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-processInfo"></a>}}
\if{latex}{\out{\hypertarget{method-processInfo}{}}}
\subsection{Method \code{processInfo()}}{
List information about the process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$processInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
