% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessGraph-class.R
\name{ProcessGraph}
\alias{ProcessGraph}
\title{Process graph}
\description{
Process graph

Process graph
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process_graph_id}}{Id of the process graph}

\item{\code{process_graph}}{Attached process graph}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProcessGraph$new()}}
\item \href{#method-load}{\code{ProcessGraph$load()}}
\item \href{#method-store}{\code{ProcessGraph$store()}}
\item \href{#method-buildExecutableProcessGraph}{\code{ProcessGraph$buildExecutableProcessGraph()}}
\item \href{#method-clone}{\code{ProcessGraph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize process graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessGraph$new(process_graph_id = NA, process_graph = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_graph_id}}{Id of the process graph}

\item{\code{process_graph}}{Attached process graph}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
Attach corresponding process graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessGraph$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
Attach process graph to the Session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessGraph$store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildExecutableProcessGraph"></a>}}
\if{latex}{\out{\hypertarget{method-buildExecutableProcessGraph}{}}}
\subsection{Method \code{buildExecutableProcessGraph()}}{
Call the private function loadProcess with given process graph and job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessGraph$buildExecutableProcessGraph(job = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job}}{Current job}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Executable process

Make an executable process for the given graph

param graph_list Graph to load the corresponding process
param job Attach the job to the process
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessGraph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
