% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session-Class.R
\name{Session}
\alias{Session}
\title{Session}
\description{
This is the session class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{processes}}{This field is also managed during runtime. Here all template processes are listed}

\item{\code{data}}{A list of products offered by the service which is managed at runtime.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Session$new()}}
\item \href{#method-getEndpoints}{\code{Session$getEndpoints()}}
\item \href{#method-getConfig}{\code{Session$getConfig()}}
\item \href{#method-startup}{\code{Session$startup()}}
\item \href{#method-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEndpoints"></a>}}
\if{latex}{\out{\hypertarget{method-getEndpoints}{}}}
\subsection{Method \code{getEndpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$getEndpoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfig"></a>}}
\if{latex}{\out{\hypertarget{method-getConfig}{}}}
\subsection{Method \code{getConfig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$getConfig()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-startup"></a>}}
\if{latex}{\out{\hypertarget{method-startup}{}}}
\subsection{Method \code{startup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$startup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
