% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session-Class.R
\name{SessionInstance}
\alias{SessionInstance}
\title{Session class}
\description{
Session class

Session class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{processes}}{Package processes}

\item{\code{data}}{Data of current SessionInstance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SessionInstance$new()}}
\item \href{#method-getEndpoints}{\code{SessionInstance$getEndpoints()}}
\item \href{#method-getConfig}{\code{SessionInstance$getConfig()}}
\item \href{#method-startSession}{\code{SessionInstance$startSession()}}
\item \href{#method-initDirectory}{\code{SessionInstance$initDirectory()}}
\item \href{#method-initEndpoints}{\code{SessionInstance$initEndpoints()}}
\item \href{#method-createEndpoint}{\code{SessionInstance$createEndpoint()}}
\item \href{#method-assignData}{\code{SessionInstance$assignData()}}
\item \href{#method-clone}{\code{SessionInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEndpoints"></a>}}
\if{latex}{\out{\hypertarget{method-getEndpoints}{}}}
\subsection{Method \code{getEndpoints()}}{
Get endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getEndpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
endpoints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfig"></a>}}
\if{latex}{\out{\hypertarget{method-getConfig}{}}}
\subsection{Method \code{getConfig()}}{
Get configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getConfig()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
configuration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-startSession"></a>}}
\if{latex}{\out{\hypertarget{method-startSession}{}}}
\subsection{Method \code{startSession()}}{
Start the session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$startSession()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initDirectory"></a>}}
\if{latex}{\out{\hypertarget{method-initDirectory}{}}}
\subsection{Method \code{initDirectory()}}{
initializes workspace and data paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$initDirectory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initEndpoints"></a>}}
\if{latex}{\out{\hypertarget{method-initEndpoints}{}}}
\subsection{Method \code{initEndpoints()}}{
biuld a df to add the endpoints later on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$initEndpoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createEndpoint"></a>}}
\if{latex}{\out{\hypertarget{method-createEndpoint}{}}}
\subsection{Method \code{createEndpoint()}}{
Create an endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$createEndpoint(path, method, handler = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path for the endpoint}

\item{\code{method}}{type of request}

\item{\code{handler}}{function to be executed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
created Endpoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignData"></a>}}
\if{latex}{\out{\hypertarget{method-assignData}{}}}
\subsection{Method \code{assignData()}}{
Function to assign data of collection to the data path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$assignData(col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{Collection of class 'Collection'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
